ipc: implement discover_root and become_root
These 2 system calls can be used to respectively discover the root
IPC node (for example the system server, which you can use to discover
other IPC services) and to become the root IPC node, which can be done
only once and will return EINVAL if you try to use it again. That
system call should be used by the system server once on startup.